Terminando las funciones

+++ src/main/java/com/practicas/model/DatosLibro.java
@@ -9,5 +9,5 @@ import java.util.List;
 public record DatosLibro(@JsonAlias("title") String titulo,
                          @JsonAlias("authors") List<DatosAutor> autor,
                          @JsonAlias("languages") List<String> idiomas,
-                         @JsonAlias("download_count") Double numeroDeDescargas) {
+                         @JsonAlias("download_count") Integer numeroDeDescargas) {
 }
